Fixed input lookup not working if you have pressed 'More input rows' (#105)
* Fixed input lookup not working if you have pressed 'More input rows'

The problem was that we now use XMLHTTPRequest and innerHTML rather than
AJAX which means it doesn't execute any scripts included in the
replacement text.

I have changed this to use createContextualFragment instead which
properly creates an element before inserting it so does run contained
scripts.

* An 'expensive' tab containing javascript would not run it.

The fix is the same as it was for `Component_Field_Action`, use
`createContextualFragment(...)` rather than `innerHTML = ...`.

* Replace some uses of `innerHTML` with `textContent`

This makes it clearer that these particular uses of innerHTML do not do
anything with html (it may also make it marginally more efficient).