Added the ai_summary ingredient to core.
This doesn't add any fields (unlike the ingredient) because it can't be
enabled by default (as it requires an endpoint) so having the fields
included by default seemed odd.

Instead it includes the function to render topics (with links to their
searches) and a comment suggesting fields to add if enabling it.