Added the ai_summary ingredient to core (#106)
* Added the ai_summary ingredient to core.

This doesn't add any fields (unlike the ingredient) because it can't be
enabled by default (as it requires an endpoint) so having the fields
included by default seemed odd.

Instead it includes the function to render topics (with links to their
searches) and a comment suggesting fields to add if enabling it.

* Replaced the comment about fields for enabled with example fields.

This would make it easier to set up in the default case as you could
just copy the commented code and un-comment it.