Refuse to add a document if the hash doesn't match with the client hash
This gets the hash passed from the browser and passes it through to
DataObj/File where it is checked against the one calculated there and
refused if it doesn't match.

NB: This currently just freezes the loading bar on 100% if the hash
doesn't match, I'm not sure if there is a better way to do this so that
we can inform the user why it failed (and whether it actually makes
sense to inform them as they can't change anything other than to try
again).