Added all of the core packages I could find us using to Bundle/EPrints
I'm undecided on whether we should have this list of core perl packages
in the Bundle however we definitely shouldn't have it with just an
arbitrary selection of the packages we actually use.

As such this adds pretty much all the packages that are included in core
(as of v5.16.1) and used somewhere in EPrints.