Prevent a javascript error when there are no search results (#107)
* Prevent a javascript error when there are no search results

When there are no search results `thisElement` is `undefined` rather
than `null` because there are no 'ep_search_result' elements. So we can
instead do a truthy check which is fine because all valid values will be
`Element`s which are not falsy.

* Prevented empty/solely ignored word searches from making invalid regex

The regex construction code was sharing a bracket unnecessarily so that
if you made an empty search (invalid to actually search but what happens
if you just click the search button) or a search involving only ignored
words (like 'if and') it would end up with mismatched brackets in the
regex.