Add a variable to disable client side hashing
This is useful because the hashing does hurt upload speeds for large
files like videos as you have to wait for it to hash before it will
start uploading to the server.

Ideally it would probably hash in a background task (web workers) while
uploading the file and then tack the hash on the end of the xhr but
there doesn't seem to be a mechanism for modifying an in-flight xhr.