Add a trigger to update the history when a subject changes
This will allow future patches to use this to show a series of
revisions.

In order to store the subject ID it has to be converted to an integer
with unpack (because history stores integer IDs) and to prevent IDs with
the same first few characters colliding it hashes it first to ensure the
whole ID is considered.

As subjects don't have connected revision files it instead stores the
details of what changed directly in the `details` field, stored as a
JSON array of arrays (so that the order is consistent).