Added diffing support for lists of values
This means the `render_history_diff` function supports all types of
values currently used in the Subject dataobj.

To display the last section of text it adds a pseudo-operation to the
end of the diff, the alternative is to have effectively a copy of the
first `render_list_portion` under the loop, but this seemed cleaner.