Fixed a bug where the 'Extract' button is instead a text box
This meant that if you uploaded a PDF (in the normal upload form) then
when you click 'Next' it would instead take you to the 'Extract' form.

The `ajax` field is used to fill in `$action{rel}` and something
relating to the bootstrap changes (specifically #119) is using that to
replace text boxes with buttons but because it didn't have that field it
was failing to convert it to a button.