Improved the ability to find Bundle/EPrints.pm
By adding not requiring a slash and adding a dot at the beginning it
will now create the correct path even if you run it with
`perl cpan_modules.pl` rather than `perl ./cpan_modules.pl`.

To help with its usability this also makes it executable by default so
that it can be run with `./cpan_modules.pl`.