Moved the History into its own file and added a tab to show it.
The tabs allows it to be out of the way unless needed and also prevents
it from loading unless requested (as it is marked with expensive). This
also allows it to be split into its own separate plugin with a screen
group of 'subject_edit_tabs'.

Aside from the changes required to make tabs work this is just a
mechanical cut-paste from 'Edit' to 'History'.