Limit the shown history fields to a fixed subset in a fixed order
Rather than showing every field (including two 'Languages' fields) it
now only shows the fields in `$c->{subject_history_fields}` which also
allows the order to be fixed to the order of items in that list.

As an implementation detail the history now stores it as a hash rather
than an array because it makes it more efficient to access in whatever
order is requested.