Fix spurious redefine warnings when running certain commands
When we call an epadmin command from another epadmin command we have to
pass `$repo` rather than `$repoid` as otherwise it will re-create the
repository and so load cfg files twice.

This showed up as an issue for `latest_additions` which defines new
`EPScript::Compiled` functions, and would therefore cause complaints
when running `generate_abstracts` and `generate_views`.