Fixes XML special characters being escaped in phrase attributes
This works by just un-escaping them in the attribute function because
<epc:print> automatically does the escaping.

Fixes #98.

It feels like it would be neater to solve this by making `print` not
escape when inside an attribute however there are a few problems with
this. The first is just that we can't tell we are in an attribute, but
this is solvable by adding a field to `params`.

The larger problem is that `print` doesn't do the escaping itself,
this is instead handled by `LibXML::createTextNode`. The proper solution
here would be to instead use `createAttribute` which doesn't do the
escaping however you can't then combine these together with `appendChild`
in the case where you have multiple `print` blocks combining to form one
attribute.