Match the json 'More input rows' button with the bootstrap buttons
By using `render_button` we can match the style of button used elsewhere
with classic_style toggled either way because the button code is now
identical.

This also makes it properly just refresh the page rather than adding the
row and exiting the form, so it is a surprise fix as well as a visual
change.