Allow script tags in search results to run when infinite scrolling
This showed up with `limit_names_shown` which adds some javascript to
hide names when there are too many, and it would just not apply this
javascript to any results past the first page.

The issue was that the new pages were being created using `innerHTML`
and added to a temporary `XMLDocument`. Because of the temporary
document even changing it to use `contextualFragment` didn't fix it
however we don't actually need a secondary document because we can pull
parts out of a document fragment just fine.