Fixed 'epadmin create' failing after setting database password
The problem was `create_db` calls `create_tables` and passes it
repository, which was changed from `$repoid` to `$repo` in 6120088
however `create_db` creates it with `db_connect => 0` which doesn't
generate the database, while `create_tables` uses the database.

The solution here is just to create the database and add it to the
repository just before calling `create_tables`.