Only display part of the documentation on an epadmin error
Rather than always displaying the whole POD documentation if a command
misses a `repo_id` or `dataset_id` this prints just the relevant parts
of the documentation. This makes it much easier to see what you have
done wrong, particularly if you are coming from the short documentation
which just shows the names of the commands.

To prevent it from matching on partially correct commands (like `epadmin
cre`) it requires the action to be followed by a space however this
should be fine as we will never have an argument error for something
without a space (as it clearly has no arguments).