Add more details to the short docs for epadmin commands
Make the short-form help more useful in `epadmin` by including details
of what arguments it expects (using the notation git uses). The goal
here is to stay within one line so that it doesn't bulk up the reply
(this is why there are other arguments that you can pass to
`create_user` that are just marked in a comment).

This is particularly important now that we show the short docs in most
cases, only showing the long docs for `help`, `--help` or `--man`.