Make 'scaleEffectOpen' work with computed styles (separate css)
Previously it would ignore any margin added by a class causing it to
stutter at the end of an expansion as it unsets height, making it
suddenly jump to full height.

This also replaces the recursive use of `setTimeout` with `setInterval`
as it is clearer to read and more efficient given that only Safari has
actually implemented tail-recursion (due to concerns about debugging).