Fixed the jolt at the start and end of 'toggleSlide'
This was caused by the padding and border of the outer element which
aren't changed by the `height` attribute being modified.

The solution was just to move the padding and border (and all the alert
styling) into the inner div which means it all gets affected by the
outer div being shrunk.