Fix the PDF import failing if there are no EPrints
Rather than using the first `EPrint` to get the workflow (and work out
whether it contains `contributions`) this instead creates a fake
`DataObj` that pretends to be an `EPrint`.

This creates a fake `DataObj` rather than a `DataObj::EPrint` because
the `get_dataset` function (the only one that matters) is simpler in
`DataObj` and only requires the `dataset` field.