Fix saved simple searches not highlighting correctly
Non-Xapian simple search is set up as a multi-field with an id `q` that
then contains all of the fields it applies to (`documents`, `abstract`
etc.). When a search happens normally it happens to place `documents` as
the first inner item which doesn't have a filter associated with it and
therefore highlights everything just fine. However when we save the
search these get re-ordered to be alphabetical and `abstract` is limited
to just the `embedded` field so it then only highlights that.

To fix this it is changed to use `get_id` rather than `get_field->name`
which will generally return the same value but for multi-fields it
doesn't just return the first item and so for simple searches it will
filter on `q` rather than `documents` or `abstract` which has no
selection filters associated with it and therefore will highlight
everything.

This is a necessarily non-perfect solution as it is still theoretically
applying filtering to simple search where it shouldn't be however there
doesn't seem to be an easy way to tell if a search is simple. One
potential improved solution would be to ignore simple search entirely
and make it correctly filter on all the fields in a `multi` (as opposed
to just the first) which would incidentally fix simple search.