Split fields into their metafields to get names when highlighting
This properly fixes simple search by highlighting all of the fields
included in the search (rather than just trying to highlight
everything). In particular it fixes embedding content into simple search
which was broken before as it considered the name to be `q`.

This currently generates rather inefficient javascript as it does the
same search multiple times but that should be easy to fix in a follow-up
commit.