Add back basic support for previews showing in the 'lightbox'
This was removed with the removal of prototypejs (4982608) although the
actual lightbox elements were retained in `00_ep_js_lib.js` (869d9a0).

The primary mechanical difference between this and the prototype
lightbox is that rather than recognising 'rel' tags that contain
'lightbox' it instead adds some javascript in `onclick`, currently this
is done directly however it will probably morph into something more akin
to `Component_Documents(...)` which associates an event listener and can
do more registering of the image (for the sake of 'Prev' and 'Next').

This initial commit adds the lightbox in a necessarily limited form with
no support for videos, 'Prev' and 'Next' buttons, and captions.