Change lightbox into a class that registers all of the click events
This will allow it to better handle `Prev` and `Next` by having a list
of all of the preview-able images on the page.

This uses `class: "lightbox"` to recognise the links (as defined by
those links available at 'DOMContentLoaded'), as opposed to
`prototypejs` which used `rel: "lightbox..."`.

By storing the image link in `href` again it allows right clicking and
selecting 'Open in new tab' to work while still correctly handling
normal clicks due to `preventDefault`. One quirk is that you can't use
Ctrl/Cmd-Click to open it in a new tab as we override the click event.