Move the 'overlay' and 'lightbox' elements creation into 'Lightbox'
The creation of the `overlay` and `lightbox` elements on window load
were moved into `00_ep_js_lib.js` in 869d9a0 as the lightbox specific
code had been deleted however now that lightbox is back it makes sense
to put them back in `40_lightbox.js`.

This also allows us to attach close event listeners to it when we create
it rather than the first time we click one of the links. This then
results in the listeners only being attached once so they don't have to
rely on event listener de-duplication and can use an anonymous
function.