Support 'Prev' and 'Next' buttons to switch between images in lightbox
This shows the `Prev` and `Next` buttons when applicable (currently just
grouping everything rather than having named groups) and makes them
correctly switch the image.

This also majorly reduces the amount of `getElementById` by storing most
of these elements on the class with `Lightbox.*`, as well as starting
the image box at a width of 660px so that it doesn't look strange during
the first load (lightbox images are 640px + 2*10px border).