Removed remaining uses of 'getElementById'
This now stores all of the elements it needs to reference on the
`Lightbox` for ease of use. It also made a tweak to how the elements are
gathered so that even if there is another item with the id `loading` for
example it will find the correct one.

Further tweaks here involve removing the check that prevents it from
initialising twice as this should never be necessary since the addition
of groups. It also removed `loadingLink` as this wasn't used and would
cause it to jump to the top of the page, and made `prevLink` and
`nextLink` show up with the select cursor rather than the text one.