Remove jQuery from 'Longtext_counter'
We don't bundle jQuery anymore and are planning on removing the
ingredient from submodules (#169) so this definitely shouldn't be using
jQuery.

One odd quirk I noticed when testing this is that if you don't have
`eprints_classic_style` it won't highlight the word counter in bold red
when you go over the word limit as the styling for `ep_over_word_limit`
is only in `form.css`.