Fix 'Longtext_counter' when used with 'multiple => 1'
Due to its use of `DOMContentLoaded` the word count would break if you
used the arrows to move the boxes (as this rebuilds the item with AJAX).
Luckily it doesn't actually need to happen once it finishes loading so
this removes that and ensures `ep_over_word_limit` is carried over as
well.