Add support for key presses to the lightbox
This adds `Esc`, `ArrowLeft` and `ArrowRight` support when the lightbox
is open where `Esc` will close the lightbox, `ArrowLeft` will move to
the previous picture and `ArrowRight` will move to the next picture.

This was a part of the old prototypejs solution however I accidentally
forgot about it in my re-creation (#184). Prototype also supported `P`
and `N` for switching between images however that feels largely
unintuitive so I haven't currently implemented it.