Prevent facets from reordering when the page is reloaded
While facets are sorted so that the more numerous terms are at the top,
any terms with the same count are ordered randomly which means that
they can switch anytime the page reloads (such as when you select a
facet).

To prevent this I've added a secondary sort for alphabetical so that iff
the count matches they will then be sorted A-Z (An annoying quirk of
this is that years will be sorted oldest to newest as that is
alphabetical).