Fix a crash when importing an EPrint with an ID'd contributor
If you create a new EPrint that already has a contributor with the
`id_value` and `id_type` fields (aka from an import) it will hit this
line which tries to pass the bare-word `type` to
`Entity::entity_with_id`.

This is because `entity_with_id` takes a hash-ref rather than a hash so
it can't be passed an un-packed hash.