Fix the Longtext_counter word limit highlighting with new styling
`Longtext_counter` uses `ep_over_word_limit` to highlight when a textbox
is over the word count limit (for example '23/20 words'). However the
highlighting for this was moved to `eprints_classic_style` in 97f65c1
with the rest of `form.css` so it fails to highlight when you go over
the word limit.