Fix a bug where the page width is used where the height is expected
This was accidentally introduced in 495b0ac where previously the
prototype-adjacent code correctly used the width and height for these
fields but it was then typoed into using the width twice here.

To make this clearer I also removed the 'getPageSizeArray' function as
the `0` and `1` are much less readable to be sure it is using the
correct value (as it wasn't).