Fix the incorrect pub_lib commit being included for ORCIDs
When adding support for importing ORCIDs with PDF metadata
(https://github.com/eprints/eprints3.5/pull/200) this involved a commit
to `flavours/pub_lib` (eprints/pub_lib@4e203bb3). When it came to
actually committing this however I adjusted its message to better
explain which main PR it relates to however forgot to adjust the commit
the main PR was pulling to eprints/pub_lib@111d678d. As such main is
currently targeting a commit which doesn't really exist on `pub_lib`
(well it probably does exist but is technically dangling) so this fixes
that by targeting the altered commit.