Prevent the 'missing archive' error showing for 'epadmin --something'
As noted in
https://github.com/eprints/eprints3.5/pull/205#pullrequestreview-3115049162
if the second argument of an epadmin command starts with something that
can't be a repository id (for example `--help`) we should suppress the
missing archive error.

Previously this would only suppress the error for 1-argument commands
even though it would only attach the correct argument if it actually
could be an archive (according to the regex). This just changes it so it
always sets `IGNORE_UNKNOWN_ARCHIVE` if it isn't passing
`--archive=...`.