Fix accessing phrases from JavaScript
This was broken by the switch to use `xhrRequest` rather than
`Ajax.Request` as `XMLHttpRequest` makes no assumptions about the data
being returned (even when it is **right there** in the 'Content-Type').
This means that all phrase requests (such as the 'Cancel' text when
uploading a document or the 'Document too large' text) would just show
up as `undefined` or '' (depending on how they are used).

The way this has been fixed in existing cases is to add a layer of JSON
parsing on top of the request however there is a mechanism for this in
`XMLHttpRequest` so it makes sense to use it. To do this we add another
optional field 'responseType' to `xhrRequest` that defaults to '' and
therefore it can be set to 'json' rather than needing to resort to
parsing it ourselves.

This also changes one existing `.response` to `.responseText` as while
it was already correct (because `.response` will default to parsing as
text) it feels best to be clear in cases when we want text (which could
theoretically allow defaulting to `responseType: 'json'`).