Remove 'bazaar' submodule
Bazaar is being entirely removed (by default) for 3.5 as ingredients now
do its job but better so this removes it from `.gitmodules` to prevent
it from being downloaded as a submodule and removes it from the flavour
ingredient lists.

This does not of course prevent bazaar from being used, it just means
that it will have to be added manually as with most other ingredients.

Fixes #90.