Fix toggling when 'show_help => always'
When `show_help` is set to 'always' for a workflow component (such as
for 'Add a new document' in the default EPrint workflow) it starts open
and so shouldn't have a button to open it. However because we weren't
correctly passing `no_toggle` into the template from the component
(which is set by `show_help`) it would just show the button and
semi-work.

To fix this we actually pass `no_toggle` into the template as it was
already reading it. This also tweaks the alt-text for the '?' button to
be a question mark rather than a plus icon as the help icon is now a
question mark.