Fix toggling when 'show_help => always'
When `show_help` is set to 'always' for a workflow component (such as
for 'Add a new document' in the default EPrint workflow) it starts open
and so the toggle button should start by shrinking it. However instead
the icon would show as '?' (rather than the '-' it should be) and the
first time you press it it re-opens the already open tab.

To fix this we check whether the help panel is collapsed and use that to
determine the initial state of the icons (and whether `toggleSlide` is
set to expand first). This also tweaks the alt-text for the '?' button
to be a question mark rather than a plus icon as the help icon is now a
question mark.