Mark default options in 'Set' fields
While the default option of a 'Set' field is frequently the same as
'selected', this isn't true in the case where you have saved a document
and then come back to edit it, as 'selected' will now be the saved value
rather than the original default.

Therefore this adds another option 'mark-default' (it's not a great name
but selected is currently called 'default') which can be used as
something that won't change when reloading a document.

This is useful when resetting these fields on the client-side as the
only thing which currently does that (contributors `unset_entity`)
manually associates the defaults with these fields and does so
incorrectly (eprints/eprints3.5#217).

This also removes the `disabled` option from `render_single_option` as
it wasn't actually used by anything and so could be cleanly replaced by
`default`.