Add clear buttons to multiple Compound fields
This fits in with the normal row-change buttons and acts akin to the
'contributors' `unset_entity` button but more generally for any
`Compound` field with `multiple => 1`.

This also uses the new `data-default` attribute to refill `select` items
with their default option if it exists, rather than always clearing them
to 'UNSPECIFIED'.

Due to how it filters for `span` elements (in order to clear the
'Entity' column) this will also accidentally remove the `unset_entity`
button however that isn't a problem as the plan is to remove this (due
to it being a pure duplication of this button).

Fixes eprints/eprints3.4#368.