Rename render_option_list 'default' parameter to 'selected'
This fits both with how it is used and how it is referred elsewhere in
the codebase. This also allows the awkwardly named 'mark-default'
parameter to be renamed to 'default' which fits how it is used much
better as well.

In order to prevent every use of this function (some of which are not in
core) from having to change, if a caller specifies 'default' without
specifying 'selected' then it will use that as the selected so it should
be backwards compatible with all previous uses.