Remove a mention to the 'CHANGELOG' file from 'NEWS'
There is no file called 'CHANGELOG' and from what I can tell `git log
--all --full-history -- "*CHANGELOG*"` it hasn't existed for a long time
(the git history only dates to the beginning of 3.4 but it didn't exist
then). As such it seems unhelpful to link to this in 'NEWS' so I have
changed it to suggest using `git log` as this will at least give you a
history (although likely both too in-depth and un-specific to be
helpful).