Still show the component title when 'show_help => "always"'
The `!no_toggle` check was being used over-eagerly and covered the title
as well as the help box (and help buttons). This shifts the check inward
so that it will only cover the show and hide buttons themselves.

This is most easily tested on the 'Upload' page which wouldn't show the
"Add a new document" header above its "To upload a document..." help
text.