Implement the 'EP_TRIGGER_LOG' trigger
While this trigger existed in `EPrints::Const` it was never actually
triggered (as warned in the docs). This changes it so where `$c->{log}`
would have been called previously it instead triggers `EP_TRIGGER_LOG`.

This replaces the old `$c->{log}` function however for backwards
compatibility it will run that function (and add a warning message) if
it still exists.