Implement the 'EP_TRIGGER_DEFAULTS' dataset trigger
While this trigger existed in `EPrints::Const` it was never actually
triggered (as warned in the docs). This adds a call to this trigger
where it would previously run the `set_*_defaults` functions.

For backwards compatibility the `set_*_defaults` functions are still run
however they log a warning suggesting upgrading to the
EP_TRIGGER_DEFAULTS trigger.

Ideally this would also update `easy_pages` however that doesn't
currently have separate 3.4 and 3.5 branches so I have created a branch
for it (https://github.com/eprints/easy_pages/tree/implement-triggers)
however haven't tracked that branch here.

NB: This includes a change to `pub_lib` and so ideally its
`implement-triggers` branch would be merged and this rebased on top of
that before merging.