Fix the title bar disappearing or showing twice
The way this template works is much more complicated than I initially
thought (hence why my tweaks broke it in strange ways). As such this
tries to reduce the repetitiveness as a whole rather than just tweaking
a line or two.

This also adds some comments to try and explain at a broad level how
everything works as it is still non-trivial and some of the variables
can be tricky  to figure out at first glance (for example `is_collapsed`
relates to whether the component as a whole is collapsible while the
`no_toggle` variable relates to whether the help box is collapsible).

Finally this cleans up a couple of little quirks like the collapsible
minus button had less margin than the plus button an had an alt text of
'+'.