Implement the 'EP_TRIGGER_VALIDATE' dataset trigger
While this trigger existed in `EPrints::Const` it was never actually
triggered (as warned in the docs). This adds a call to this trigger
where it would previously run the `validate_*` functions.

For backwards compatibility the `validate_*` functions are still run
however they log a warning suggesting upgrading to the
EP_TRIGGER_VALIDATE trigger.

NB: This includes a change to `pub_lib` so ideally its
`trigger-functions` branch would be merged and this rebased on top of
that before merging.