Fix toggling when 'show_help => always' (#216)
When `show_help` is set to 'always' for a workflow component (such as
for 'Add a new document' in the default EPrint workflow) it starts open
and so shouldn't have a button to open it. However because we weren't
correctly passing `no_toggle` into the template from the component
(which is set by `show_help`) it would just show the button and
semi-work.

This simplifies the template for this and fixes the particular
interaction of `no_toggle` that caused this.