Allow collapsible items with a value still collapse
If you add some text to a collapsible component on an EPrint and then
save and reload the form it would prevent you from shrinking that
component again.

This was caused by the `MetaField->can_collapse` function which checks
whether it currently contains a value (or for `Multi` whether any of the
fields contains a value). This is then read by `Component->is_collapsed`
so to check whether a field is collapsible (irrespective) of whether it
is currently collapsed we can just read `Component->{collapse}`.

This then controls which elements are set to `ep_no_js` (aka hidden off
the bat) and the '+' and '-' buttons are informed whether they start
visible or not.

Fixes #222.