Prevent opening a collapsible component from stuttering (#223)
This was previously fixed in #148 for help boxes however it still
happened for collapsible elements (like 'Contact Email Address').

It was caused by the `scaleEffectOpen` not correctly handling elements
with padding (as it sets the height but padding is handled separately)
and so was fixed by just shifting the padding in a layer so that the div
that we expand has no padding and therefore its height includes the
padding of the internal element.