Allow the 'clear' button to clear more field types (#225)
Longtext elements use `textarea` not `input` so they would not correctly
clear when pressing the 'clear button' (#218). It also has a slightly
different ID pattern for some reason.

Richtext fields were also not being cleared as they use an embedded
iframe that then updates a textarea (unreliably) in the background.
While it is possible to directly clear this iframe it seems much more
sane to use the `tinymce` methods to do this instead.

This also handles an edge case around the 'longtext counter' field as
this uses the 'input' event which is only triggered automatically by
_user_ input. This would mean that the word count would remain as `x/20`
until you enter some new text into the field.