Make clicking the description of a radio button count as the button (#228)
The convention is for radio buttons to be controllable by the text next
to them so this uses `label`s tracking the `input`s.

This is a 3.5 version of eprints/eprints3.4#520 as it requires some
slightly layout tweaks that are different in 3.4 (notably the addition
of `pe-1` to the header element).

Fixes #227.