Remove the dependency on `latest_tool` from the `zero` flavour
The `zero` flavour (in this case `lib/.../index.xpage`) should have
minimal dependencies possible so therefore shouldn't require the
`latest_additions` ingredient to work. Previously this was fine as
`cgi/latest_tool` also existed in core, however this makes limited sense
at this point so is being removed.

The alternative would be to leave `cgi/latest_tool` in but strip it down
to the bare minimum (just enought to handle `output=RSS`) however it
doesn't seem unreasonable that to do anything related to the most recent
publications you need the ingredient designed around that (and it
doesn't prevent all RSS feeds, just these ones).