Track the main branches of submodules rather than the feature branches
The latest_additions/remove-35-latest-tool and
pub_lib/latest-additions-margin branches have been merged into their
respective main branches so this updates it to track the merge commits.

This also includes a minor rebased change to `pub_lib` where I noticed
that I hadn't correctly specified the ingredient dependency as it has to
be either `ingredient:` or `ingredients/` not `ingredients:`.