Use a CSS transition for the document 'Show options'
Previously this used `EPJS_toggleSlideScroll` and a pair of
`EPJS_toggle` which this replaces with `EPJS_checkboxSlide` and some
CSS.

The main advantage of this is simplicity, `EPJS_toggleSlideScroll` does
some fairly nasty stuff with repeatedly setting the height in a
`setInterval` whereas instead we can just set the height once and rely
on a CSS transition (courtesy of `ep_toggleable`) to do the animation.

This does sadly still need javascript as there is no way to expand to or
from `height: auto` in pure CSS (except in an extremely hacky way on
very modern Chromium).

This is also based on a `checkbox` rather than a pair of input images so
that whether it is showing the `+ Show options` or `- Hide options` is
just based on a little bit of CSS and the current status of the (hidden)
checkbox.

One final benefit of the new system is that it correctly listens for
`prefers-reduced-motion` although that could be edited into the old
system (via `matchMedia`).