Switch component 'help' and 'collapse' to use `EPJS_checkboxSlide`
As with the earlier commits, this reduces the amount of manual
javascript height editing in favour of CSS. In this case it also makes
the HTML structure considerably more sane by making the title bar always
show and including both the + and - buttons on it (for collapsible).

This does unfortunately include a Baseline 2023 bit of CSS (`:has`)
however if the user doesn't have that feature it will just show the help
button even when the component is collapsed which is not a major issue.