Prevent reordering richtext components from clearing their contents
This is a slightly hacky fix but it checks how many items are in the
`multiple` component and then loops through them all, saving each one.

This will also attempt to do the same thing when re-ordering
non-richtext fields but `tinymce.get` will just return `null` causing it
to be ignored.

Fixes #226, #229.