Remove `latest_tool` in favour of the `latest_additions` ingredient (#230)
This is effectively identical (particularly as of
eprints/latest_additions#16) and the ingredient is included by default
on 3.5 now.

This also pulls a group of recent `latest_additions` commits and a
commit to `pub_lib` as `latest_tool` had a special bit of margin to
handle the front page better.

The `zero` flavour (in this case `lib/.../index.xpage`) should have
minimal dependencies possible so therefore shouldn't require the
`latest_additions` ingredient to work. Previously this was fine as
`cgi/latest_tool` also existed in core, however this makes limited sense
at this point so is being removed.