Implement the 'EP_TRIGGER_WARNINGS' dataset trigger
While this trigger existed in `EPrints::Const` it was never actually
triggered (as warned in the docs). This adds a call to this trigger
where it would previously run the `*_warnings` functions.

For backwards compatibility the `*_warnings` functions are still run
however they log a warning suggesting upgrading to the
EP_TRIGGER_WARNINGS trigger.

NB: This includes a change to `pub_lib` so ideally its
`trigger-functions` branch would be merged and this rebased on top of
that before merging.