Match the 'Edit page' hyperlink to look like the others (#240)
* Match the 'Edit page' hyperlink to look like the others

As this button only shows up conditionally (you need to have an
`index.xpage` file copied into your archive) it was missed from the
original set of changes for menu buttons.

This changes the class to the same as the other buttons in that menu,
making it centered and black.

Fixes #235.

* Use same approach as Admin/Config/Phrases.pm for action_render_link (#235)

---------

Co-authored-by: David R Newman <drn@ecs.soton.ac.uk>