Convert `EPrints::Box` to expand with CSS transitions (#239)
* Convert `EPrints::Box` to expand with CSS transitions

This was introduced in #232 however that left `EPrints::Box` and
`ajax/subject_input` alone as I couldn't find an easy way to test them.

I have since become aware that the Help tab (for example on the 'Manage
Deposits' page) is a box (added via `epv:box`) so have switched this
over to use `checkboxSlide`.

As with #232, this improves the sanity of the HTML by only hiding and
showing the '?' and '-' icons and doesn't have to repeatedly set height
in a loop with javascript.

The only thing still using `EPJS_toggleSlideScroll` in core is
`ajax/subject_input` however there is no great reason to switch that
over as it should be left for backwards compatibility anyway.

* Fixes wrong alt text for show_icon_url

---------

Co-authored-by: David R Newman <drn@ecs.soton.ac.uk>